Fixing duration dep, transformer issue
The ruby-duration dependency has a dependency on activesupport, which is
fairly loose, however, in version 4.2.0 of activesupport they removed
something the duration gem was using, so we're staying at the 3.1
version.

The transformer now correctly checks for the existence of an image
before doing changes on it. It also logs issues if it finds any in the
form of standard awestruct errors.