Add ability for submodules in .git or module dir
We've found depending on the version the submodule .git directory can be
in either <submodule>/.git or <root>/.git/module. This should work
regardless of which way the submodules are created.

Also removed an .orig file that shouldn't have been there.

Trying new idea to find git directory

Removing puts